refactor(tegg): migrate agent-runtime types to @eggjs/tegg-types
Move type definitions (AgentStore, ObjectStorageClient, errors) from
@eggjs/agent-runtime to @eggjs/tegg-types/agent-runtime for centralized
type management. Implementation code (OSSAgentStore, OSSObjectStorageClient,
AgentStoreUtils) stays in agent-runtime. The agent-runtime package re-exports
all types from tegg-types for backward compatibility.
