+++ tegg/core/agent-runtime/src/index.ts
@@ -1,8 +1,6 @@
-export * from './AgentStore.ts';
-export * from './errors.ts';
-export * from './FileAgentStore.ts';
-export * from './MessageConverter.ts';
-export * from './ObjectStorageClient.ts';
+// Re-export types from @eggjs/tegg-types (backward compatible)
+export * from '@eggjs/tegg-types/agent-runtime';
+// Implementation code
 export * from './OSSObjectStorageClient.ts';
 export * from './OSSAgentStore.ts';
 export * from './AgentStoreUtils.ts';
